Refactor: move storage related types into types/v065. Dir "types" includes all data-compatibility related types
