+++ openraft/src/core/append_entries.rs
@@ -3,18 +3,18 @@ use crate::core::RaftCore;
 use crate::core::State;
 use crate::core::UpdateCurrentLeader;
 use crate::error::RaftResult;
-use crate::raft::AppendEntriesRequest;
-use crate::raft::AppendEntriesResponse;
-use crate::raft::Entry;
-use crate::raft::EntryPayload;
-use crate::AppData;
-use crate::AppDataResponse;
-use crate::EffectiveMembership;
-use crate::LogId;
+use crate::types::v065::AppData;
+use crate::types::v065::AppDataResponse;
+use crate::types::v065::AppendEntriesRequest;
+use crate::types::v065::AppendEntriesResponse;
+use crate::types::v065::EffectiveMembership;
+use crate::types::v065::Entry;
+use crate::types::v065::EntryPayload;
+use crate::types::v065::LogId;
+use crate::types::v065::RaftNetwork;
+use crate::types::v065::RaftStorage;
 use crate::MessageSummary;
 use crate::RaftError;
-use crate::RaftNetwork;
-use crate::RaftStorage;
 use crate::Update;
 
 impl<D: AppData, R: AppDataResponse, N: RaftNetwork<D>, S: RaftStorage<D, R>> RaftCore<D, R, N, S> {
