+++ openraft/src/core/replication.rs
@@ -16,14 +16,14 @@ use crate::raft::RaftRespTx;
 use crate::replication::RaftEvent;
 use crate::replication::ReplicaEvent;
 use crate::replication::ReplicationStream;
-use crate::storage::Snapshot;
 use crate::summary::MessageSummary;
-use crate::AppData;
-use crate::AppDataResponse;
-use crate::LogId;
-use crate::NodeId;
-use crate::RaftNetwork;
-use crate::RaftStorage;
+use crate::types::v065::AppData;
+use crate::types::v065::AppDataResponse;
+use crate::types::v065::LogId;
+use crate::types::v065::NodeId;
+use crate::types::v065::RaftNetwork;
+use crate::types::v065::RaftStorage;
+use crate::types::v065::Snapshot;
 use crate::ReplicationMetrics;
 
 impl<'a, D: AppData, R: AppDataResponse, N: RaftNetwork<D>, S: RaftStorage<D, R>> LeaderState<'a, D, R, N, S> {
