+++ openraft/src/error.rs
@@ -4,10 +4,10 @@ use std::collections::BTreeSet;
 use std::fmt::Debug;
 use std::time::Duration;
 
-use crate::raft::Membership;
-use crate::raft_types::SnapshotSegmentId;
-use crate::LogId;
-use crate::NodeId;
+use crate::types::v065::LogId;
+use crate::types::v065::Membership;
+use crate::types::v065::NodeId;
+use crate::SnapshotSegmentId;
 use crate::StorageError;
 
 /// A result type where the error variant is always a `RaftError`.
